New Development: No             JIRA Issue: Yes (CAS-3749/TRAC-266)         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs: interactive tests are required I. on CASA (1) plot something by CASA sdplot task (2) run CASA sdfit with plotlevel!=0 (3) close plotter window by pressing X button on the window (4) plot something by sdplot task again. II. on stand-alone ASAP (1) plot something by asapplotter (2) run asapfitter.auto_fit with plot=T (3) plot something by asapplotter again
  Put in Release Notes: No
             Module(s): sdplot, asapplotter
           Description: 
Fixed a bug which caused sdplot and asapplotter crash after closing
plotter loaded by the other modules (tasks) with X button. It was
a bug caused by the fact ID numbers of figure managers are 
identical in both sdplot (asapplotter) and the other plotter. 
It is also possible, that user loads plotter with the ID identical 
to asapplotter and overrides it. This is because, user can select 
arbitrary ID when generating a new plot.

asap.plotter._assert_plotter does more tests to make sure the 
previous plotter GUI is alive, and properly reloads plotter if not.

Also some minor fixes to reset linewidth for plot invoked by
modules other than asapplot/sdplot, i.e., scantable, asapfitter,
and asapmath.