New Development: No             JIRA Issue: Yes CAS-4105         Ready for Test: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: sdmath unit test   Put in Release Notes: No              Module(s): Module Names change impacts.            Description: Describe your changes here...
Unified behavior of scantable operation with scalar, 1-d array, 
and 2-d array. Result of the operation (+,-,*,/) is returned as 
new scantable. Returned scantable refers new table if sd.rcParams['insitu'] 
is False while it refers the same table as input scantable if 
sd.rcParams['insitu'] is True. For the later case, input scantable will 
be modified. If sd.rcParams['scantable.storage'] is 'disk', it means 
that original scantable (on disk) will be modified so that those 
functions must be used with care.

In task level (sdmath and sdscale), we take care to keep original table 
unchanged so the users don't need to worry about that as long as they 
are working on task level.