New Development: No             JIRA Issue: Yes (CAS-4859)         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs: set asap.rcParams[scantable.storage] = 'disk' run asap.plotter.plot exit from casapy or asap. A temporary scantable on disk should be deleted.
  Put in Release Notes: No
             Module(s): asap.plotter, sdplot
           Description: 
There were cyclic references between plotters and toolbars, i.e., plotter and CustomToolbarCommon,
and flagplotter and CustomFlagToolbarCommon, which prevented deletion of temporary scantables
from disk when exitting sessions.
I changed the reference to plotters in toolbars as a weak reference so that cyclic referencing
should not avoid desctruction of the plotters. It should be ok because there shouldn't be a case
where toolbar exists without plotter, but I'm not 100% sure to be honest.